Create bounding capsules
However, this method is not open to rotate with current configuration as the position of each capsule is located in the middle of each bone.
This can be resolved by re-ordering the process (rotate bones first then create bounding capsule from there), however this requires more computing power as it has to set the position and rotation per frame.